Move printing of system info out of StelLogger
